requirements.txt
openai==1.12.0
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.2.0
tabulate==0.9.0
